renaming face processing module file
